requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0
matplotlib>=3.7.0
customtkinter>=5.2.0
pillow>=10.0.0
pyserial>=3.5
